requirements.txt
playwright
spotipy
rich
python-dotenv
curl_cffi
beautifulsoup4
pyjson5
